Add NodeInfo object for algorithm.send.
This object is meant mostly to be an argument to algorithm.send, but also
as a generall wrapper for the identity of a node in the network.

Algorithm.send still needs to be available with an IP argument aswell. This
is needed for replying to DTOs. There are 2 solutions: make DTO's contain
NodeInfo objects or making a function that neatly returns a nodeInfo to which
one can reply.